Minor refactoring of constraints. Case insensitive ExactConstraint comparisons now ignore accents
